Add files via upload
In today's fast-paced world, managing a bakery efficiently is crucial for success. 
The Bakery Management System is designed to streamline the operations of a 
bakery, making it easier for owners and staff to handle various tasks effectively. 
This comprehensive system leverages the power of Python to create a user friendly and feature-rich solution for bakery management